Changed the type of user_id in the sessions sql table to be bigint so I could use uint as the type for it in Go.

+++ models/sql/sessions.sql
@@ -1,6 +1,6 @@
 create table sessions (
   id serial primary key,
-  user_id int unique,
+  user_id bigint unique,
   token_hash text unique not null,
   foreign key (user_id) references users(id)
 );
